+++ packages/patrol_cli/lib/src/test_bundler.dart
@@ -32,7 +32,6 @@ import 'dart:async';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:patrol/patrol.dart';
-import 'package:patrol/src/platform/platform_automator.dart';
 import 'package:patrol/src/platform/contracts/contracts.dart';
 import 'package:test_api/src/backend/invoker.dart';
 
@@ -73,7 +72,9 @@ Future<void> main() async {
   // Dart test (out of which they had been created) and wait for it to complete.
   // The result of running the Dart test is the result of the native test case.
 
-  final platformAutomator = PlatformAutomator(config: PlatformAutomatorConfig.defaultConfig());
+  final platformAutomator = PlatformAutomator(
+    config: PlatformAutomatorConfig.defaultConfig(),
+  );
   await platformAutomator.initialize();
   final binding = PatrolBinding.ensureInitialized(platformAutomator);
   final testExplorationCompleter = Completer<DartGroupEntry>();
@@ -87,7 +88,8 @@ Future<void> main() async {
     // Maybe somewhat counterintuitively, this callback runs *after* the calls
     // to group() below.
     final topLevelGroup = Invoker.current!.liveTest.groups.first;
-    final dartTestGroup = createDartTestGroup(topLevelGroup,
+    final dartTestGroup = createDartTestGroup(
+      topLevelGroup,
       tags: ${tags != null ? "'$tags'" : null},
       excludeTags: ${excludeTags != null ? "'$excludeTags'" : null},
     );
@@ -140,14 +142,15 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:patrol/patrol.dart';
 import 'package:patrol/src/platform/contracts/contracts.dart';
-import 'package:patrol/src/platform/platform_automator.dart';
 
 // START: GENERATED TEST IMPORTS
 ${generateImports(testDirectory, [testFilePath])}
 // END: GENERATED TEST IMPORTS
 
 Future<void> main() async {
-  final platformAutomator = PlatformAutomator(config: PlatformAutomatorConfig.defaultConfig());
+  final platformAutomator = PlatformAutomator(
+    config: PlatformAutomatorConfig.defaultConfig(),
+  );
   await platformAutomator.initialize();
   
   PatrolBinding.ensureInitialized(platformAutomator)