+++ packages/patrol/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Add `PlatformAutomator` to exports. (#2833)
+
 ## 4.0.1
 
 - Remove dependency on `js`. (#2831)